Fix permission error interaction between 2.6.2 and upstream
Changes meant that permissions were too broad in `home/opuser/.op`.
My initial attempt would have fixed the problem during existing
deployments, but would not have been viable in the Helm chart.
While the permissions would have been valid for the `.op` folder,
1Password SCIM Bridge would not have been able to save files to the `.op`
directory, resulting in a similar issue as prior to this fix.
This fix solves the problem by keeping the permissions for `.op` as
in the current chart, and then sets the `umask` for `.op` to ensure that
newly created files have the appropriate permissions.

Resolves !3088.