Lots of broken stuff
I think the best way forward that eliminates the most repeated code is to have a bool for .Values.scim.credentials.workspaceUser which, if true, checks for Workspace secrets, and if false only checks for scimsession file. This may make it easier to populate stringData of secret.yaml template appropriately.